refactor: extract shared token-metrics code into lib modules (DRY)
