+++ src/admin/lib/session-scanner.ts
@@ -1,0 +1,233 @@
+/**
+ * @module session-scanner
+ *
+ * Shared session file scanning logic for the token metrics pipeline.
+ *
+ * Scans OpenClaw gateway transcripts and Claude Code session logs,
+ * extracting usage records within a [fromMs, cutoffMs) time range,
+ * and merging them into hourly buckets.
+ *
+ * Shared by collect-token-metrics and recalculate-token-metrics.
+ */
+
+import fs from 'node:fs';
+import path from 'node:path';
+
+import { SESSIONS_DIR } from '../../lib/constants.js';
+import type { CursorState, HourlyBucket } from '../types/token-metrics.js';
+import { mergeUsage, tsToHour } from './bucket-io.js';
+import { detectChannel, registerChannelName } from './channel-mapper.js';
+import { listCCSessionFiles, parseCCLine } from './claude-code-scanner.js';
+import { normalizeUsage, parseUsageLine } from './usage-parser.js';
+
+/** Result returned by scanAllSessions. */
+export interface ScanResult {
+  buckets: Map<string, HourlyBucket>;
+  seenModels: Set<string>;
+  ocProcessed: number;
+  ocSkipped: number;
+  ccProcessed: number;
+  ccSkipped: number;
+}
+
+/**
+ * Process a single OpenClaw session file: extract usage records
+ * within [fromMs, cutoffMs), detect channel, and merge into buckets.
+ */
+function processOCFile(
+  filePath: string,
+  fileName: string,
+  cursors: CursorState,
+  fromMs: number,
+  cutoffMs: number,
+  buckets: Map<string, HourlyBucket>,
+  seenModels: Set<string>,
+): void {
+  const stat = fs.statSync(filePath);
+  const cursor = cursors[fileName] as CursorState[string] | undefined;
+  const byteOffset = cursor?.byteOffset ?? 0;
+
+  // Skip if fully processed and file hasn't grown
+  if (byteOffset >= stat.size) return;
+
+  const content = fs.readFileSync(filePath, 'utf8');
+  const allLines = content.split('\n');
+
+  // Detect channel from the first few lines (always, even if resuming)
+  // Use first 50 lines for channel detection. 15 was too narrow — long-lived
+  // topic thread sessions often have metadata/custom lines filling the first 15,
+  // pushing recognizable patterns (heartbeat, Slack context) just out of reach.
+  const channelResult = detectChannel(allLines.slice(0, 50));
+  if (channelResult.key.startsWith('slack:channel:')) {
+    const name = channelResult.name;
+    if (name.startsWith('#')) {
+      registerChannelName(channelResult.key, name);
+    }
+  }
+
+  // Compute byte positions per line to know where to resume
+  let bytePos = 0;
+  let maxProcessedTs = cursor?.lastTimestamp ?? 0;
+
+  for (const line of allLines) {
+    const lineByteLen = Buffer.byteLength(line, 'utf8') + 1; // +1 for \n
+    const lineStart = bytePos;
+    bytePos += lineByteLen;
+
+    // Skip lines we've already processed
+    if (lineStart < byteOffset) continue;
+    if (!line.trim()) continue;
+
+    const parsed = parseUsageLine(line);
+    if (!parsed) continue;
+
+    // Only include records within range
+    if (parsed.tsMs < fromMs) continue;
+    // Skip messages in the current (incomplete) hour
+    if (parsed.tsMs >= cutoffMs) continue;
+
+    const modelKey = [parsed.provider, parsed.model].join('/');
+    seenModels.add(modelKey);
+    const usage = normalizeUsage(parsed.rawUsage, modelKey);
+    const hour = tsToHour(parsed.tsMs);
+
+    mergeUsage(buckets, hour, channelResult.key, modelKey, usage);
+
+    if (parsed.tsMs > maxProcessedTs) {
+      maxProcessedTs = parsed.tsMs;
+    }
+  }
+
+  // Update cursor: advance byteOffset to current file size
+  cursors[fileName] = {
+    byteOffset: stat.size,
+    lastTimestamp: maxProcessedTs,
+  };
+}
+
+/**
+ * Scan all OpenClaw and Claude Code session files, extracting usage
+ * records within [fromMs, cutoffMs) and merging into hourly buckets.
+ *
+ * For the regular collector, pass fromMs=0 (no lower bound).
+ * For recalculation, pass the user's fromMs.
+ */
+export function scanAllSessions(
+  fromMs: number,
+  cutoffMs: number,
+  cursors: CursorState,
+  ccCursors: CursorState,
+): ScanResult {
+  const buckets = new Map<string, HourlyBucket>();
+  const seenModels = new Set<string>();
+
+  // ── OpenClaw sessions ──
+  const allFiles = fs.readdirSync(SESSIONS_DIR);
+  const sessionFiles = allFiles.filter(
+    (f) =>
+      f.endsWith('.jsonl') ||
+      f.includes('.jsonl.deleted.') ||
+      f.includes('.jsonl.reset.'),
+  );
+
+  let ocProcessed = 0;
+  let ocSkipped = 0;
+
+  for (const fileName of sessionFiles) {
+    const filePath = path.join(SESSIONS_DIR, fileName);
+    const cursor = cursors[fileName] as CursorState[string] | undefined;
+    const stat = fs.statSync(filePath);
+
+    if (cursor?.byteOffset !== undefined && cursor.byteOffset >= stat.size) {
+      ocSkipped++;
+      continue;
+    }
+
+    processOCFile(
+      filePath,
+      fileName,
+      cursors,
+      fromMs,
+      cutoffMs,
+      buckets,
+      seenModels,
+    );
+    ocProcessed++;
+  }
+
+  // ── Claude Code sessions ──
+  const ccFiles = listCCSessionFiles();
+  let ccProcessed = 0;
+  let ccSkipped = 0;
+
+  for (const ccFile of ccFiles) {
+    let stat: fs.Stats;
+    try {
+      stat = fs.statSync(ccFile.filePath);
+    } catch {
+      continue;
+    }
+
+    const cursor = ccCursors[ccFile.cursorKey] as
+      | CursorState[string]
+      | undefined;
+    if (cursor?.byteOffset !== undefined && cursor.byteOffset >= stat.size) {
+      ccSkipped++;
+      continue;
+    }
+
+    const byteOffset = cursor?.byteOffset ?? 0;
+    const content = fs.readFileSync(ccFile.filePath, 'utf8');
+    const allLines = content.split('\n');
+
+    let bytePos = 0;
+    let maxProcessedTs = cursor?.lastTimestamp ?? 0;
+
+    for (const line of allLines) {
+      const lineByteLen = Buffer.byteLength(line, 'utf8') + 1;
+      const lineStart = bytePos;
+      bytePos += lineByteLen;
+
+      if (lineStart < byteOffset) continue;
+      if (!line.trim()) continue;
+
+      const record = parseCCLine(line);
+      if (!record) continue;
+
+      if (record.tsMs < fromMs) continue;
+      if (record.tsMs >= cutoffMs) continue;
+
+      seenModels.add(record.modelKey);
+      const usage = normalizeUsage(
+        {
+          input: record.usage.input,
+          output: record.usage.output,
+          cacheRead: record.usage.cacheRead,
+          cacheWrite: record.usage.cacheWrite,
+        },
+        record.modelKey,
+      );
+      const hour = tsToHour(record.tsMs);
+      mergeUsage(buckets, hour, ccFile.channelKey, record.modelKey, usage);
+
+      if (record.tsMs > maxProcessedTs) {
+        maxProcessedTs = record.tsMs;
+      }
+    }
+
+    ccCursors[ccFile.cursorKey] = {
+      byteOffset: stat.size,
+      lastTimestamp: maxProcessedTs,
+    };
+    ccProcessed++;
+  }
+
+  return {
+    buckets,
+    seenModels,
+    ocProcessed,
+    ocSkipped,
+    ccProcessed,
+    ccSkipped,
+  };
+}
